# packages/runtime-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/openwa-safety/workspace/quiescence": {
      "get": {
        "operationId": "OpenWASafetyController_workspaceQuiescence",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenWASafetyQuiescenceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Check whether the managed workspace has drained all in-flight OpenWA work",
        "tags": [
          "openwa-safety"
        ]
      }
    },
    "/api/v1/openwa-safety/workspace/control": {
      "post": {
        "operationId": "OpenWASafetyController_workspaceControl",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWASafetyControlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenWASafetyScopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Block or resume every OpenWA operation in the managed workspace",
        "tags": [
          "openwa-safety"
        ]
      }
    },
    "/api/v1/openwa-safety/sessions/{sessionId}": {
      "get": {
        "operationId": "OpenWASafetyController_snapshot",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenWASafetyScopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read the durable OpenWA safety state for one session",
        "tags": [
          "openwa-safety"
        ]
      }
    },
    "/api/v1/openwa-safety/sessions/{sessionId}/quiescence": {
      "get": {
        "operationId": "OpenWASafetyController_quiescence",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenWASafetyQuiescenceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Check whether one session has drained all in-flight OpenWA work",
        "tags": [
          "openwa-safety"
        ]
      }
    },
    "/api/v1/openwa-safety/sessions/{sessionId}/control": {
      "post": {
        "operationId": "OpenWASafetyController_control",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWASafetyControlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenWASafetyScopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Manually block or resume OpenWA operations for one session",
        "tags": [
          "openwa-safety"
        ]
      }
    },
    "/api/v1/openwa-safety/sessions/{sessionId}/profile": {
      "put": {
        "operationId": "OpenWASafetyController_profile",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWASafetyProfileChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenWASafetyScopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Change the enforced OpenWA safety profile for one session",
        "tags": [
          "openwa-safety"
        ]
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "SessionController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List allowlisted Gateway sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{id}": {
      "get": {
        "operationId": "SessionController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read a Gateway session from the durable read model",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{id}/sync": {
      "post": {
        "operationId": "SessionController_requestSync",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          },
          "409": {
            "description": "The operation key conflicts or another sync mode is active",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/RuntimeErrorDto"
                    },
                    {
                      "$ref": "#/components/schemas/SyncModeConflictDto"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Queue durable discovery and group reconciliation",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{id}/sync-runs/{runId}": {
      "get": {
        "operationId": "SessionController_getSyncRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read sync progress",
        "tags": [
          "sessions"
        ]
      }
    },
...
  },
  "info": {
    "title": "WA Runtime API",
    "description": "Stable, versioned API contract consumed by WA Studio",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "runtime-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Runtime-Key"
      }
    },
    "schemas": {
      "OpenWASafetyQuiescenceDto": {
        "type": "object",
        "properties": {
          "drained": {
            "type": "boolean"
          },
          "processingMessageJobs": {
            "type": "number"
          },
          "unsettledConnectorCommands": {
            "type": "number"
          },
          "activeSafetyLeases": {
            "type": "number"
          },
          "checkedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "drained",
          "processingMessageJobs",
          "unsettledConnectorCommands",
          "activeSafetyLeases",
          "checkedAt"
        ]
      },
      "OpenWASafetyControlDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "BLOCK",
              "RESUME"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "action"
        ]
      },
      "OpenWASafetyScopeDto": {
        "type": "object",
        "properties": {
          "scopeType": {
            "type": "string",
            "enum": [
              "WORKSPACE",
              "UPSTREAM",
              "SESSION"
            ]
          },
          "effectiveScopeType": {
            "type": "string",
            "enum": [
              "WORKSPACE",
              "UPSTREAM",
              "SESSION"
            ]
          },
          "circuitState": {
            "type": "string",
            "enum": [
              "CLOSED",
              "OPEN",
              "HALF_OPEN",
              "MANUAL_BLOCKED"
            ]
          },
          "rateMode": {
            "type": "string",
            "enum": [
              "NORMAL",
              "THROTTLED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "THROTTLED",
              "COOLDOWN",
              "RECOVERY",
              "BLOCKED"
            ]
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "cooldownUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "profile": {
            "type": "string",
            "enum": [
              "CANARY",
              "STANDARD"
            ]
          },
          "policyVersion": {
            "type": "number"
          },
          "revision": {
            "type": "number"
          },
          "lastSuccessAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastFailureAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "scopeType",
          "effectiveScopeType",
          "circuitState",
          "rateMode",
          "status",
          "reason",
          "cooldownUntil",
          "profile",
          "policyVersion",
          "revision",
          "lastSuccessAt",
          "lastFailureAt",
          "updatedAt"
        ]
      },
      "OpenWASafetyProfileChangeDto": {
        "type": "object",
        "properties": {
          "profile": {
            "type": "string",
            "enum": [
              "CANARY",
              "STANDARD"
            ]
          }
        },
        "required": [
          "profile"
        ]
      },
      "RuntimeErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "CAMPAIGN_NOT_FOUND"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "SyncModeConflictDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "code": {
            "type": "string",
            "example": "SYNC_MODE_CONFLICT"
          },
          "message": {
            "type": "string",
            "example": "A different synchronization mode is already active"
          },
          "activeRunId": {
            "type": "string",
            "format": "uuid"
          },
          "activeMode": {
            "type": "string",
            "enum": [
              "FULL",
              "INCREMENTAL"
            ]
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "activeRunId",
          "activeMode"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "ready"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "pushName": {
            "type": "string",
            "nullable": true
          },
          "connectedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastActiveAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "engineLoaded": {
            "type": "boolean"
          },
          "lastError": {
            "type": "string",
            "nullable": true
          },
          "restriction": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "gatewayCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "gatewayUpdatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "syncedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "phone",
          "pushName",
          "connectedAt",
          "lastActiveAt",
          "engineLoaded",
          "lastError",
          "restriction",
          "gatewayCreatedAt",
          "gatewayUpdatedAt",
          "syncedAt"
        ]
      },
      "SessionListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "SyncRequestDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "FULL",
              "INCREMENTAL"
            ],
            "default": "FULL",
            "description": "FULL reconciles every discovered group; INCREMENTAL reconciles only new, changed, invalidated or stale groups. Omission preserves FULL behavior."
          }
        }
      },
      "SyncRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "syncType": {
            "type": "string",
            "enum": [
              "FULL",
              "INCREMENTAL"
            ]
          },
          "phase": {
            "type": "string",
            "enum": [
              "DISCOVERING",
              "RECONCILING",
              "COMPLETED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "groupsSynced": {
            "type": "number"
          },
          "groupsDiscovered": {
            "type": "number",
            "description": "Total groups returned by authoritative discovery"
          },
          "groupsScheduled": {
            "type": "number",
            "description": "Groups selected for detail reconciliation in this run"
          },
          "groupsFailed": {
            "type": "number",
            "description": "Groups whose reconciliation exhausted retries"
          },
          "groupsSkipped": {
            "type": "number",
            "description": "Groups skipped because they disappeared during reconciliation"
          },
          "groupsPending": {
            "type": "number",
            "description": "Group reconciliations waiting for their first attempt"
          },
          "groupsRunning": {
            "type": "number",
            "description": "Group reconciliations that currently own a processing lease"
          },
          "groupsRetrying": {
            "type": "number",
            "description": "Group reconciliations waiting for a durable retry"
          },
          "membersSynced": {
            "type": "number",
            "description": "Members observed in successfully reconciled group snapshots; not rows changed"
          },
          "nextAttemptAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cooldownUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sessionId",
          "syncType",
          "phase",
          "status",
          "groupsSynced",
          "groupsDiscovered",
          "groupsScheduled",
          "groupsFailed",
          "groupsSkipped",
          "groupsPending",
          "groupsRunning",
          "groupsRetrying",
          "membersSynced",
          "nextAttemptAt",
          "cooldownUntil",
          "error",
          "requestedAt",
          "startedAt",
          "completedAt"
        ]
      },
...
    }
  }
}